# swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: SaltRiver
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - multipart/form-data
  - application/x-www-form-urlencoded
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /relation:
    x-swagger-router-controller: friend
    get:
      tags:
        - friend
      description: 관계 등록 한 데이터에서 전체 받아오기
      operationId: getRelations
      security:
        - jwt: []
      parameters:
        - name: search_text
          in: query
          description: search_text
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - friend
      description: 신규 관계 등록시 작동하는 api
      operationId: addRelations
      security:
        - jwt: []
      parameters:
        - name: id
          in: body
          description: friend_id
          schema:
            type: string
            description : friend_id
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - friend
      description: 관계 상태 업데이트
      operationId: setRelation
      summary: 관계 상태 업데이트
      security:
        - jwt: []
      parameters:
        - name: relation
          in: body
          description: CorpCode(send relation offer)
          schema:
            type: object
            properties:
              idx:
                type: integer
                description: 상태를 변경하고자 하는 관계의 인덱스
              stat:
                type: string
                description: 변경하고자 하는 상태 (요청/승인/거절)
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - friend
      description: Delete relation API
      operationId: deleteRelation
      security:
        - jwt: []
      parameters:
        - name: idx
          in: query
          description: Index that we wanted to delete
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /file:
    x-swagger-router-controller: file
    post:
      operationId: addimageFile
      tags:
        - upload
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: file
          description: The file to upload
          type: file
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /file/{fileName}:
    x-swagger-router-controller: file
    get:
      operationId: getImageFile
      tags:
        - upload
      produces:
        - image
        - image/png
        - image/gif
        - image/jpeg
      parameters:
        - name: fileName
          in: path
          description: file name
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: file
...
securityDefinitions:
  jwt:
    type: apiKey
    in: header
    name: jwt
# complex objects have schema definitions
definitions:
...
  SuccessResponse:
    properties:
      data:
        type: object
        properties:
          success:
            type: boolean
            description: 성공 여부
          message:
            type: string
            description: 메시지
  RelationResponse:
    properties:
      data:
        type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string